src/requirements.txt
aiofiles==24.1.0
aiofirebase==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomqtt==2.3.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==24.3.0
backoff==2.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
firebase_admin==6.5.0
firebaseaio==0.1.1
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gcloud-aio-auth==5.4.1
gcloud-aio-storage==9.4.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-genai==1.19.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.5.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
msgpack==1.0.8
multidict==6.1.0
oauth2client==4.1.3
paho-mqtt==2.1.0
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.3
pvporcupine==3.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyBluez==0.30
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydbus==0.6.0
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.0
SpeechRecognition==3.10.4
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==15.0.1
yarl==1.18.3
